pyproject.toml
[project]
name = "rasp_rest_api"
version = "0.1.0"
description = ""
authors = [
    {name = "SelfTopic",email = "chestor.official@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "pillow (>=11.3.0,<12.0.0)"
]
package-mode = false
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
